src/pyproject.toml
[project]
name = "mcp-workshop"
version = "0.1.0"
description = "Codebase for MCP Workshop of Hong Kong Python User Group"
authors = [
    {name = "wyhwong"},
    {name = "alex-au-922"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "fastmcp (>=2.5.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
